# Server/StreamMoviesServer/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Genre:
    properties:
      genre_id:
        type: integer
      genre_name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - genre_id
    - genre_name
    type: object
...
  models.Movie:
    properties:
      _id:
        type: string
      admin_review:
        type: string
      genre:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      imdb_id:
        type: string
      poster_path:
        type: string
      ranking:
        $ref: '#/definitions/models.Ranking'
      title:
        maxLength: 500
        minLength: 2
        type: string
      youtube_id:
        type: string
    required:
    - genre
    - imdb_id
    - poster_path
    - ranking
    - title
    - youtube_id
    type: object
...
  models.Ranking:
    properties:
      ranking_name:
        type: string
      ranking_value:
        type: integer
    required:
    - ranking_name
    - ranking_value
    type: object
  models.UpdateReview:
    properties:
      admin_review:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a movie streaming application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Movie Stream API
  version: "1.0"
paths:
  /addmovie:
    post:
      consumes:
      - application/json
      description: Add a new movie to the database
      parameters:
      - description: Movie object
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/models.Movie'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add a movie
      tags:
      - movies
  /genre:
    post:
      consumes:
      - application/json
      description: Add a new genre or update an existing one
      parameters:
      - description: Genre object
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/models.Genre'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add or update a genre
      tags:
      - genres
  /genres:
    get:
      consumes:
      - application/json
      description: Get a list of all movie genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all genres
      tags:
      - genres
...
  /movie/{imdb_id}:
    get:
      consumes:
      - application/json
      description: Get a single movie by its IMDB ID
      parameters:
      - description: IMDB ID
        in: path
        name: imdb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a movie by ID
      tags:
      - movies
  /movie/{imdb_id}/updatereview:
    patch:
      consumes:
      - application/json
      description: Update the admin review and ranking of a movie
      parameters:
      - description: IMDB ID
        in: path
        name: imdb_id
        required: true
        type: string
      - description: Review object
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.UpdateReview'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a movie review
      tags:
      - movies
  /movies:
    get:
      consumes:
      - application/json
      description: Get a list of all movies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all movies
      tags:
      - movies
  /recommendedmovies:
    get:
      consumes:
      - application/json
      description: Get a list of recommended movies for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Movie'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get recommended movies
      tags:
      - movies
...
swagger: "2.0"